Executer: Add interpreter tracing mode for method tracing facility.
The AcpiDebugTrace() is the mechanism known as ACPI method tracing that is
used by Linux as ACPICA debugging message reducer. This facility can be
controlled through Linux ACPI subsystem - /sys/module/acpi/parameters.
This facility requires CONFIG_ACPI_DEBUG to be enabled to see ACPICA trace
logs in the kernel dmesg output.

This patch enhances AcpiDebugTrace() to make it not only a message reducer,
but a real tracer to trace AML interpreter execution. Note that in addition
to the AML tracer enabling, this patch also updates the facility with the
following enhancements:
1. Allow a full path to be specified by the AcpiDebugTrace() API.
2. Allow any method rather than just the entrance of AcpiEvaluateObject()
   to be traced.
3. All interpreter ACPI_LV_TRACE_POINT messages are collected for
   ACPI_EXECUTER layer.

This patch also updates "trace" command to allow it to perform all Linux
method tracing functionalities:
   1. "trace aml" to always enable AML execution trace messages
      (level=TRACE_POINT, layer=EXECUTER) for all control methods.
   2. "trace aml \MTHD" to always enable AML execution trace messages for
      only \MTHD method.
   3. "trace aml \MTHD once" to enable AML execution trace messages only
      once for only \MTHD method.
   4. "trace enable" to always enable all trace messsages (level=ALL,
      layer=ALL) for all control methods.
   5. "trace enable \MTHD" to always enable all trace messages for only
      \MTHD method.
   6. "trace enable \MTHD once" to enable all trace messages only once for
      only \MTHD method.
   7. "trace disable" to disable trace facility.

The Makefiles of acpinames/iasl are also updated to include exdebug.o and
the duplicated stubs are removed after that.

Note that since this patch has enhanced the method tracing facility, Linux
need also be updated after applying this patch. Lv Zheng.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>