Dispatcher: Cleanup ACPI_OPERAND_OBJECT's AML address assignments.
This patch cleans up the code of assigning the AML address to the
ACPI_OPERAND_OBJECT.

The idea behind this cleanup is:
The AML address of the ACPI_OPERAND_OBJECT should always be determined at
the point where the object is encountered. It should be started from the
first byte of the object. For example, the opcode of the object, the name
string of the UserTerm object, or the first byte of the packaged object
(where a PkgLength is prefixed). So it's not cleaner to have it assigned
here and there in the entire ACPICA source tree.

There are some special cases for the internal opcodes, before cleaning up
the internal opcodes, we should also determine the rules for the AML
addresses of the internal opcodes:
1. INT_NAMEPATH_OP: the address of the first byte for the NameString.
2. INT_METHODCALL_OP: the address of the first byte for the NameString.
3. INT_BYTELIST_OP: the address of the first byte for the ByteData list.
4. INT_EVAL_SUBTREE_OP: the address of the first byte for the
                        Region/Package/Buffer/BankField/Field arguments.
5. INT_NAMEDFIELD_OP: the address to the NameSeg.
6. INT_RESERVEDFIELD_OP: the address to the 0x00 prefix.
7. INT_ACCESSFIELD_OP: the address to the 0x01 prefix.
8. INT_CONNECTION_OP: the address to the 0x02 prefix.
9: INT_EXTACCESSFIELD_OP: the address to the 0x03 prefix.
10.INT_RETURN_VALUE_OP: the address of the replaced operand.
11.ComputationalData: the address to the
                      Byte/Word/Dword/Qword/StringPrefix.

Before cleaning up the internal root scope of the AmlWalk, turning it into
the TermList, we need to remember the AmlStart address as the "Aml"
attribute for the ACPI_OPERAND_OBJECT created by AcpiPsCreateScopeOp().

Finally, we can delete some redundant AML address assignment in psloop.c.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>