disassembler: print warning message when emitting conflicting external declarations when using AcpiDmEmitExternals.
Conflicting external declarations are illegal in ASL. When the
disassembler encounters a conflicting external in the AML, the
conflicting external is emitted with a comment warning the user
that a conflicting external has been emitted. The alternative
approach to handling conflicting external declarations within
AML is for the disassembler to abort. This approach is not
preferred because the purpose of the disassembler is to
disassemble the AML rather than prevent the disassembly due to
the quality of AML.

Signed-off-by: Erik Schmauss <erik.schmauss@intel.com>