ASLTS: add acpiexec sanity test
The acpiexec sanity check compares the namespace output of acpiexec with an previous
know good version of the namespace output from acpiexec. Because ASLTS is written in
ASL, there may be some errors that could be unnoticed. For example, the following
ASL code may not be detected if the string object creation mechanism breaks:

Name (TXT1, "abcd")
if (TXT1 != "abcd")
{
	err(....)
}

However, if we keep a copy of old acpiexec outputs we have a regression point. By
comparing the acpiexec outputs themselves, we can perform a simple sanity check to make
sure that acpiexec outputs are the same between versions. For the example above, if
there has been a bug introduced to change the value of TXT1 to be "efgh", we can compare
this with an output to see if the value of TXT1 is "abcd".

The steps of the sanity check are as follows:
  1.) load all acpi tables from tests/aslts/src/sanity/tables on acpiexec
  2.) dump the namespace and save the output to tempOutput.txt
  3.) Compare this with the existing goodOutput.txt

Help options to update goodOutput are available. ./aslts.sh -h lists options for the sanity
test.

Signed-off-by: Erik Schmauss <erik.schmauss@intel.com>